requirements.txt
Flask==2.3.2
flask-cors==3.0.10
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.24.3
